requirements.txt
python-dotenv
llama-index
llama-parse
llama-index-llms-openai
llama-index-embeddings-openai
llama-index-vector-stores-chroma
llama-index-vector-stores-qdrant
llama_index-readers-web
llama-index-postprocessor-flag-embedding-reranker
FlagEmbedding
spacy
nest-asyncio
chromadb
llama-index-core
pandas
qdrant-client
pydantic
llama-index-callbacks-arize-phoenix
arize-phoenix[evals]
ipython
llama-index-multi-modal-llms-openai
llama-index-vector-stores-qdrant
llama_index
ftfy
regex
tqdm
llama-index-embeddings-clip
matplotlib
scikit-image
pillow